# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: LandIt API
  description: LandIt 智能求职助手 API 接口文档
  version: 1.0.0
  contact:
    name: Azir
servers:
  - url: http://localhost:8080
    description: 开发环境
...
paths:
...
  /api/user/profile:
    get:
      tags:
        - user
      summary: 获取当前用户信息
      operationId: getUserProfile
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_User'
    put:
      tags:
        - user
      summary: 更新用户信息
      operationId: updateUserProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_User'
  /api/user/avatar:
    post:
      tags:
        - user
      summary: 上传头像
      operationId: uploadAvatar
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_AvatarUpload'

  # ==================== 简历模块 ====================
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        avatar:
          type: string
          nullable: true
        email:
          type: string
        phone:
          type: string
        targetPosition:
          type: string
        targetSalary:
          type: string
        experience:
          type: string
        education:
          type: string
        skills:
          type: array
          items:
            type: string
        createdAt:
          type: string
    UserUpdateRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        targetPosition:
          type: string
        targetSalary:
          type: string
        skills:
          type: array
          items:
            type: string
    AvatarUploadResponse:
      type: object
      properties:
        avatarUrl:
          type: string

    # ==================== 简历相关 ====================
...
    ApiResponse_User:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/User'
        timestamp:
          type: integer
          format: int64
    ApiResponse_AvatarUpload:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/AvatarUploadResponse'
        timestamp:
          type: integer
          format: int64
...